Use our own babel-eslint fork, with the reference to patch-eslint-scope removed, and: (#2001)
- remove unneeded code from webpack config
- fix PR build URL to use https
- fix the a11y linting bug
- extract jest messages to common
