+++ .eslintrc
@@ -43,7 +43,10 @@
     "no-param-reassign": ["error", { "props": false }],
     "camelcase": "error",
     "react-hooks/rules-of-hooks": "error",
-    "react-hooks/exhaustive-deps": "warn"
+    "react-hooks/exhaustive-deps": "warn",
+    "jsx-a11y/href-no-hash": "off",
+    "jsx-a11y/anchor-is-valid": ["warn", { "aspects": ["noHref", "invalidHref"] }],
+    "jsx-a11y/label-has-for": "off"
   },
   "overrides": [
     {
