+++ package.json
@@ -54,7 +54,7 @@
   "devDependencies": {
     "@types/react": "^16.8.12",
     "all-contributors-cli": "^5.4.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "codesandbox/babel-eslint#v10.0.2",
     "eslint": "5.16.0",
     "eslint-config-airbnb": "^15.0.1",
     "eslint-config-prettier": "^4.2.0",
@@ -62,7 +62,7 @@
     "eslint-loader": "^1.7.1",
     "eslint-plugin-flowtype": "^2.34.0",
     "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-jsx-a11y": "^6.1.0",
     "eslint-plugin-react": "~7.4.0",
     "eslint-plugin-react-hooks": "^1.6.0",
     "flow-bin": "^0.72.0",
