+++ packages/app/src/app/components/CodeEditor/Monaco/workers/linter/index.js
@@ -30,7 +30,7 @@ const allRules = {
   'jsx-a11y/aria-role': require('eslint-plugin-jsx-a11y/lib/rules/aria-role'),
   'jsx-a11y/aria-unsupported-elements': require('eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements'),
   'jsx-a11y/heading-has-content': require('eslint-plugin-jsx-a11y/lib/rules/heading-has-content'),
-  'jsx-a11y/href-no-hash': require('eslint-plugin-jsx-a11y/lib/rules/href-no-hash'),
+  'jsx-a11y/anchor-is-valid': require('eslint-plugin-jsx-a11y/lib/rules/anchor-is-valid'),
   'jsx-a11y/iframe-has-title': require('eslint-plugin-jsx-a11y/lib/rules/iframe-has-title'),
   'jsx-a11y/img-redundant-alt': require('eslint-plugin-jsx-a11y/lib/rules/img-redundant-alt'),
   'jsx-a11y/no-access-key': require('eslint-plugin-jsx-a11y/lib/rules/no-access-key'),
@@ -286,13 +286,18 @@ const DEFAULT_RULES = {
   'jsx-a11y/accessible-emoji': 'warn',
   'jsx-a11y/alt-text': 'warn',
   'jsx-a11y/anchor-has-content': 'warn',
+  'jsx-a11y/anchor-is-valid': [
+    'warn',
+    {
+      aspects: ['noHref', 'invalidHref'],
+    },
+  ],
   'jsx-a11y/aria-activedescendant-has-tabindex': 'warn',
   'jsx-a11y/aria-props': 'warn',
   'jsx-a11y/aria-proptypes': 'warn',
   'jsx-a11y/aria-role': 'warn',
   'jsx-a11y/aria-unsupported-elements': 'warn',
   'jsx-a11y/heading-has-content': 'warn',
-  'jsx-a11y/href-no-hash': 'warn',
   'jsx-a11y/iframe-has-title': 'warn',
   'jsx-a11y/img-redundant-alt': 'warn',
   'jsx-a11y/no-access-key': 'warn',
