+++ packages/app/src/sandbox/eval/tests/jest-lite.ts
@@ -1,5 +1,8 @@
 import { dispatch, actions, listen } from 'codesandbox-api';
 import { react, reactTs } from '@codesandbox/common/lib/templates';
+import { messages } from '@codesandbox/common/lib/utils/jest-lite';
+export { messages };
+
 import expect from 'jest-matchers';
 import jestMock from 'jest-mock';
 import jestTestHooks from 'jest-circus';
@@ -85,20 +88,6 @@ function resetTestState() {
   });
 
   setState(INITIAL_STATE);
-}
-
-export enum messages {
-  INITIALIZE = 'initialize_tests',
-  ADD_FILE = 'add_file',
-  REMOVE_FILE = 'remove_file',
-  FILE_ERROR = 'file_error',
-  TOTAL_TEST_START = 'total_test_start',
-  TOTAL_TEST_END = 'total_test_end',
-  TEST_START = 'test_start',
-  TEST_END = 'test_end',
-  DESCRIBE_START = 'describe_start',
-  DESCRIBE_END = 'describe_end',
-  ADD_TEST = 'add_test',
 }
 
 export default class TestRunner {
