+++ packages/common/src/utils/jest-lite.ts
@@ -1,0 +1,13 @@
+export enum messages {
+  INITIALIZE = 'initialize_tests',
+  ADD_FILE = 'add_file',
+  REMOVE_FILE = 'remove_file',
+  FILE_ERROR = 'file_error',
+  TOTAL_TEST_START = 'total_test_start',
+  TOTAL_TEST_END = 'total_test_end',
+  TEST_START = 'test_start',
+  TEST_END = 'test_end',
+  DESCRIBE_START = 'describe_start',
+  DESCRIBE_END = 'describe_end',
+  ADD_TEST = 'add_test',
+}
